chore: handle db is down
