+++ src/clients/index.ts
@@ -1,8 +1,8 @@
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient } from '@prisma/client';
 
-export * from "./AxelarClient";
-export * from "./GMPListenerClient";
-export * from './SigningClient'
-export * from './IbcClient'
-export * from './EvmClient'
+export * from './AxelarClient';
+export * from './GMPListenerClient';
+export * from './SigningClient';
+export * from './IbcClient';
+export * from './EvmClient';
 export const prisma = new PrismaClient();
